requirements.txt
langchain==0.1.20
langchain-groq==0.1.4
langchain-community==0.0.38
chromadb==0.4.24
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.1
click==8.1.7
sentence-transformers==2.2.2